readded fetch_tweet constants to constants/index.js, commented out delete action in message.js, changed profilecard component to procard container in Homepage.js, moved postbar from components to containers, fixed import in reducers/messages.js
