+++ src/actions/constants/index.js
@@ -22,3 +22,6 @@ export const FETCH_TWEETS_FAILURE = "FETCH_TWEETS_FAILURE";
 export const FETCH_USERS_BEGIN = "FETCH_USERS_BEGIN";
 export const FETCH_USERS_SUCCESS = "FETCH_USERS_SUCCESS";
 export const FETCH_USERS_FAILURE = "FETCH_USERS_FAILURE";
+
+export const CREATE_TWEET = 'CREATE_TWEET'
+export const DELETE_TWEET = 'DELETE_TWEET'
