+++ src/actions/messages.js
@@ -2,11 +2,13 @@ import {
   FETCH_TWEETS_FAILURE,
   FETCH_TWEETS_SUCCESS,
   FETCH_TWEETS_BEGIN,
+  CREATE_TWEET,
+  //DELETE_TWEET,
   domain,
   handleJsonResponse,
   jsonHeaders
-} from "./constants";
-import {store} from "../index"
+} from "../actions/constants";
+import { store } from "../index";
 
 export const fetchTweetsBegin = () => ({
   type: FETCH_TWEETS_BEGIN
@@ -42,31 +44,43 @@ function handleErrors(response) {
   return response;
 }
 
-
-export const CREATE_TWEET = 'CREATE_TWEET'
-export const DELETE_TWEET = 'DELETE_TWEET'
-
 export const createTweet = text => dispatch => {
-  const token = store.getState().auth.login.token
+  const token = store.getState().auth.login.token;
   return fetch(domain + "/messages", {
     method: "POST",
     headers: {
       ...jsonHeaders,
       Authorization: `Bearer ${token}`
     },
-    body: JSON.stringify({text})
+    body: JSON.stringify({ text })
   })
     .then(handleJsonResponse)
     .then(res => {
-      console.log(res)
+      console.log(res);
       return dispatch({
         type: CREATE_TWEET,
         payload: res.messsage
-      })
-    })
-  // return { type: CREATE_TWEET, text}
-}
+      });
+    });
+};
 
-export const deleteTweet = id => dispatch => {
-  // return {type: DELETE_TWEET, id}
-}
+// export const deleteTweet = id => dispatch => {
+//   const token = store.getState().auth.login.token
+//   return fetch(domain + "/messages", {
+//     method: "DELETE",
+//     headers: {
+//       ...jsonHeaders,
+//       Authorization: `Bearer ${token}`
+//     },
+//     body: JSON.stringify({text})
+//   })
+//     .then(handleJsonResponse)
+//     .then(res => {
+//       console.log(res)
+//       return dispatch({
+//         type: DELETE_TWEET,
+//         payload: res.messsage
+//       })
+//     })
+//   // return {type: DELETE_TWEET, id}
+// }
