+++ src/components/Homepage.js
@@ -1,10 +1,11 @@
 import React, { Component } from "react";
 import NavBarComp from "./NavBarComp";
 import ProfileCard from "./ProfileCard";
-import Postbar from "./Postbar";
+import Postbar from "../containers/Postbar";
 import { Container, Row, Col } from "shards-react";
 import VisibleFeed from "../containers/VisibleFeed";
 import Users from "../containers/Users";
+import ProCard from "../containers/ProCard";
 
 export default class Homepage extends Component {
   render() {
@@ -14,14 +15,14 @@ export default class Homepage extends Component {
         <Container>
           <Row>
             <Col>
-              <ProfileCard />
+              <ProCard/>
             </Col>
             <Col>
                 <Row>
                     <Postbar />
                 </Row>
                 <Row>
-                    {<VisibleFeed />}
+                    <VisibleFeed />
                 </Row>
             </Col>
             <Col>
